requirements.txt
django==1.8
django-haystack
elasticsearch
requests
djangorestframework